+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## Changelog
 
+##### 1.2.1
+
+- format code for react-autocomplete-textarea lib
+
 ##### 1.2.0
 
-- Format react-autocomplete-textarea lib
+- completed react-autocomplete-textarea lib
 
 ##### 1.0.0